# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Painless GrowthOS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /v1/contacts/{id}:
    get:
      operationId: getContact
      tags: [contacts]
      parameters:
        - $ref: "#/components/parameters/IdPathParam"
      responses:
        "200":
          description: Contact
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    patch:
      operationId: updateContact
      tags: [contacts]
      parameters:
        - $ref: "#/components/parameters/IdPathParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    delete:
      operationId: deleteContact
      tags: [contacts]
      parameters:
        - $ref: "#/components/parameters/IdPathParam"
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
  /v1/leads:
    get:
      operationId: listLeads
      tags: [leads]
      parameters:
        - name: status
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/LeadStatus"
        - name: assignedUserId
          in: query
          required: false
          schema:
            type: string
        - name: source
          in: query
          required: false
          description: Filter by lead source value (e.g. 'nationwide-inquiry').
          schema:
            type: string
        - name: search
          in: query
          required: false
          description: Case-insensitive match on the contact name or service type.
          schema:
            type: string
        - name: limit
          in: query
          required: false
          description: Max rows to return (server caps at 200).
          schema:
            type: integer
        - name: offset
          in: query
          required: false
          description: Number of rows to skip, for paging past earlier results (server treats negatives as 0).
          schema:
            type: integer
        - name: hasUnreadPortalMessage
          in: query
          required: false
          description: When true, return only leads where the most recent portal message is newer than the most recent team reply.
          schema:
            type: boolean
      responses:
        "200":
          description: Leads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lead"
    post:
      operationId: createLead
      tags: [leads]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
  /v1/leads/bulk:
    post:
      operationId: bulkUpdateLeads
      tags: [leads]
      summary: Apply a bulk action (stage change, assign, tag) to many leads
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkLeadActionRequest"
      responses:
        "200":
          description: Bulk action result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkLeadActionResult"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
  /v1/leads/{id}/merge:
    post:
      operationId: mergeLead
      tags: [leads]
      summary: Merge a duplicate lead into this lead — activities and tags move to the survivor, the source lead is marked lost
      parameters:
        - $ref: "#/components/parameters/IdPathParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MergeLeadRequest"
      responses:
        "200":
          description: Surviving lead after merge
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
        "400":
          description: Invalid merge (e.g. merging a lead into itself)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Surviving or source lead not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /v1/saved-filters:
    get:
      operationId: listSavedFilters
      tags: [leads]
      summary: List the current user's saved filters
      responses:
        "200":
          description: Saved filters
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SavedFilter"
    post:
      operationId: createSavedFilter
      tags: [leads]
      summary: Save a filter preset for the current user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SavedFilterInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SavedFilter"
        "400":
          description: Invalid saved filter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /v1/saved-filters/{id}:
    delete:
      operationId: deleteSavedFilter
      tags: [leads]
      summary: Delete one of the current user's saved filters
      parameters:
        - $ref: "#/components/parameters/IdPathParam"
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /v1/leads/{id}:
    get:
      operationId: getLead
      tags: [leads]
      parameters:
        - $ref: "#/components/parameters/IdPathParam"
      responses:
        "200":
          description: Lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    patch:
      operationId: updateLead
      tags: [leads]
      parameters:
        - $ref: "#/components/parameters/IdPathParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /v1/leads/{id}/behavior:
    get:
      operationId: getLeadBehavior
      tags: [leads]
      description: Attribution details plus the linked visitor's website behavior summary. Anonymous visitors are never linked, so unlinked leads return an empty behavior shape.
      parameters:
        - $ref: "#/components/parameters/IdPathParam"
      responses:
        "200":
          description: Behavior summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeadBehaviorSummary"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
  /v1/leads/{id}/activities:
    get:
      operationId: listLeadActivities
      tags: [activities]
      parameters:
        - $ref: "#/components/parameters/IdPathParam"
      responses:
        "200":
          description: Activity timeline for a lead
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Activity"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    post:
      operationId: createLeadActivity
      tags: [activities]
      parameters:
        - $ref: "#/components/parameters/IdPathParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActivityInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Activity"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /v1/leads/{id}/send-email:
    post:
      operationId: sendLeadEmail
      tags: [leads]
      summary: Compose and send an email to the lead's contact via the connected Gmail account
      parameters:
        - $ref: "#/components/parameters/IdPathParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadSendEmailInput"
      responses:
        "201":
          description: Email sent; returns the timeline activity that logged it
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Activity"
        "400":
          description: Invalid email or the lead's contact has no email address
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Lead not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "502":
          description: Gmail send failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "503":
          description: Gmail is not connected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /v1/leads/{id}/photos/request-url:
    post:
      operationId: requestLeadPhotoUploadUrl
      tags: [leads]
      summary: Request a presigned upload URL for a rep-uploaded lead photo
      description: Returns a short-lived presigned PUT URL and objectPath. The client PUTs the file directly to storage, then calls POST /leads/{id}/photos to register the path.
      parameters:
        - $ref: "#/components/parameters/IdPathParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadUrlRequest"
      responses:
        "200":
          description: Presigned URL ready for direct upload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadUrlResponse"
        "400":
          description: Invalid file type or oversized file
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: Caller lacks crm.write permission
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Lead not found or belongs to a different org
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
  /v1/leads/{id}/conversations:
    get:
      operationId: listLeadConversations
      tags: [conversations]
      summary: List AI concierge conversations (with transcripts) for a lead
      parameters:
        - $ref: "#/components/parameters/IdPathParam"
      responses:
        "200":
          description: Conversations linked to the lead
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ConciergeConversation"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
  /v1/leads/{id}/next-action:
    get:
      operationId: getLeadNextAction
      tags: [playbooks]
      summary: Next best action for a lead, with an AI draft when it's a message
      parameters:
        - $ref: "#/components/parameters/IdPathParam"
      responses:
        "200":
          description: The recommended next action
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NextBestAction"
        "404":
          description: Lead not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
components:
  parameters:
    IdPathParam:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    Contact:
      type: object
      required: [id, organizationId, firstName, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        organizationId:
          type: string
          format: uuid
        firstName:
          type: string
        lastName:
          type: [string, "null"]
        email:
          type: [string, "null"]
        phone:
          type: [string, "null"]
        preferredContactMethod:
          type: [string, "null"]
        doNotContact:
          type: boolean
          description: Hard do-not-contact flag; blocks ALL automated outreach on every channel.
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    ContactUpdate:
      type: object
      properties:
        firstName:
          type: string
          minLength: 1
        lastName:
          type: [string, "null"]
        email:
          type: [string, "null"]
        phone:
          type: [string, "null"]
        preferredContactMethod:
          type: [string, "null"]
        doNotContact:
          type: boolean
...